Move FlatpakExport to its own file
This is pure code motion
