+++ common/Makefile.am.inc
@@ -43,6 +43,8 @@ libflatpak_common_la_SOURCES = \
 	common/flatpak-run.h \
 	common/flatpak-context.c \
 	common/flatpak-context.h \
+	common/flatpak-exports.c \
+	common/flatpak-exports.h \
 	common/flatpak-portal-error.c \
 	common/flatpak-portal-error.h \
 	common/flatpak-utils.c \
